+++ src/GitHub_Updater/Bitbucket_Enterprise_API.php
@@ -18,14 +18,23 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 /**
- * Class Bitbucket_API
+ * Class Bitbucket_Enterprise_API
  *
- * Get remote data from a Bitbucket repo.
+ * Get remote data from a self-hosted Bitbucket Server repo.
+ * Assumes an owner == project_key
  *
  * @package Fragen\GitHub_Updater
  * @author  Andy Fragen
+ * @author  Bjorn Wijers
  */
-class Bitbucket_API extends API {
+class Bitbucket_Enterprise_API extends Bitbucket_API {
+
+	/**
+	 * Holds loose class method name.
+	 *
+	 * @var null
+	 */
+	private static $method = null;
 
 	/**
 	 * Constructor.
@@ -38,11 +47,11 @@ class Bitbucket_API extends API {
 
 		$this->load_hooks();
 
-		if ( ! isset( self::$options['bitbucket_username'] ) ) {
-			self::$options['bitbucket_username'] = null;
+		if ( ! isset( self::$options['bitbucket_enterprise_username'] ) ) {
+			self::$options['bitbucket_enterprise_username'] = null;
 		}
-		if ( ! isset( self::$options['bitbucket_password'] ) ) {
-			self::$options['bitbucket_password'] = null;
+		if ( ! isset( self::$options['bitbucket_enterprise_password'] ) ) {
+			self::$options['bitbucket_enterprise_password'] = null;
 		}
 		add_site_option( 'github_updater', self::$options );
 	}
@@ -75,10 +84,13 @@ class Bitbucket_API extends API {
 		$response = isset( $this->response[ $file ] ) ? $this->response[ $file ] : false;
 
 		if ( ! $response ) {
-			$response = $this->api( '/1.0/repositories/:owner/:repo/src/:branch/' . $file );
+			self::$method = 'file';
+			$path         = '/1.0/projects/:owner/repos/:repo/browse/' . $file;
 
-			if ( $response && isset( $response->data ) ) {
-				$contents = $response->data;
+			$response = $this->api( $path );
+
+			if ( $response ) {
+				$contents = $this->bbenterprise_recombine_response( $response );
 				$response = $this->get_file_headers( $contents, $this->type->type );
 				$this->set_repo_cache( $file, $response );
 			}
@@ -104,10 +116,12 @@ class Bitbucket_API extends API {
 		$response  = isset( $this->response['tags'] ) ? $this->response['tags'] : false;
 
 		if ( ! $response ) {
-			$response = $this->api( '/1.0/repositories/:owner/:repo/tags' );
-			$arr_resp = (array) $response;
+			$response = $this->api( '/1.0/projects/:owner/repos/:repo/tags' );
 
-			if ( ! $response || ! $arr_resp ) {
+			if ( ! $response ||
+			     ( isset( $response->size ) && $response->size < 1 ) ||
+			     isset( $response->errors )
+			) {
 				$response          = new \stdClass();
 				$response->message = 'No tags found';
 			}
@@ -152,7 +166,8 @@ class Bitbucket_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/1.0/repositories/:owner/:repo/src/:branch/' . $changes );
+			self::$method = 'changes';
+			$response     = $this->bbenterprise_fetch_raw_file( $changes );
 
 			if ( ! $response ) {
 				$response          = new \stdClass();
@@ -160,6 +175,7 @@ class Bitbucket_API extends API {
 			}
 
 			if ( $response ) {
+				$response = wp_remote_retrieve_body( $response );
 				$response = $this->parse_changelog_response( $response );
 				$this->set_repo_cache( 'changes', $response );
 			}
@@ -203,11 +219,17 @@ class Bitbucket_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/1.0/repositories/:owner/:repo/src/:branch/' . 'readme.txt' );
+			self::$method = 'readme';
+			$response     = $this->bbenterprise_fetch_raw_file( 'readme.txt' );
 
 			if ( ! $response ) {
 				$response          = new \stdClass();
 				$response->message = 'No readme found';
+			}
+
+			if ( $response ) {
+				$response = wp_remote_retrieve_body( $response );
+				$response = $this->parse_readme_response( $response );
 			}
 		}
 
@@ -235,8 +257,12 @@ class Bitbucket_API extends API {
 	public function get_repo_meta() {
 		$response = isset( $this->response['meta'] ) ? $this->response['meta'] : false;
 
+		if ( $this->exit_no_update( $response ) ) {
+			return false;
+		}
+
 		if ( ! $response ) {
-			$response = $this->api( '/2.0/repositories/:owner/:repo' );
+			$response = $this->api( '/1.0/projects/:owner/repos/:repo' );
 
 			if ( $response ) {
 				$response = $this->parse_meta_response( $response );
@@ -268,10 +294,10 @@ class Bitbucket_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/1.0/repositories/:owner/:repo/branches' );
-
-			if ( $response ) {
-				foreach ( $response as $branch => $api_response ) {
+			$response = $this->api( '/1.0/projects/:owner/repos/:repo/branches' );
+			if ( $response->values ) {
+				foreach ( $response->values as $value ) {
+					$branch              = $value->displayId;
 					$branches[ $branch ] = $this->construct_download_link( false, $branch );
 				}
 				$this->type->branches = $branches;
@@ -299,65 +325,69 @@ class Bitbucket_API extends API {
 	 * @return string $endpoint
 	 */
 	public function construct_download_link( $rollback = false, $branch_switch = false ) {
-
-		if ( ! empty( $this->type->enterprise_api ) ) {
-			$download_link_base = implode( '/', array(
-				$this->type->enterprise_api,
-				$this->type->owner,
-				'repos',
-				$this->type->repo,
-			) );
-		} else {
-			$download_link_base = implode( '/', array(
-				'https://bitbucket.org',
-				$this->type->owner,
-				$this->type->repo,
-				'get/',
-			) );
-		}
-
-		$endpoint = '';
-
-		if ( $this->type->release_asset && '0.0.0' !== $this->type->newest_tag ) {
-			return $this->make_release_asset_download_link();
-		}
-
-		/*
-		 * Check for rollback.
+		/**
+		 * Downloads requires the official stash-archive plugin which enables
+		 * subdirectory support using the prefix query argument
+		 *
+		 * @link https://bitbucket.org/atlassian/stash-archive
 		 */
-		if ( ! empty( $_GET['rollback'] ) &&
-		     ( isset( $_GET['action'] ) && 'upgrade-theme' === $_GET['action'] ) &&
-		     ( isset( $_GET['theme'] ) && $this->type->repo === $_GET['theme'] )
-		) {
-			$endpoint .= $rollback . '.zip';
+		$download_link_base = implode( '/', array(
+			$this->type->enterprise,
+			'rest/archive/1.0/projects',
+			$this->type->owner,
+			'repos',
+			$this->type->repo,
+			'archive',
+		) );
 
-			// for users wanting to update against branch other than master or not using tags, else use newest_tag
-		} elseif ( 'master' != $this->type->branch || empty( $this->type->tags ) ) {
-			if ( ! empty( $this->type->enterprise_api ) ) {
-				$endpoint = add_query_arg( 'at', $this->type->branch, $endpoint );
-			} else {
-				$endpoint .= $this->type->branch . '.zip';
-			}
-		} else {
-			if ( ! empty( $this->type->enterprise_api ) ) {
-				$endpoint = add_query_arg( 'at', $this->type->newest_tag, $endpoint );
-			} else {
-				$endpoint .= $this->type->newest_tag . '.zip';
-			}
-		}
+		self::$method = 'download_link';
+		$endpoint     = $this->add_endpoints( $this, '' );
 
-		/*
-		 * Create endpoint for branch switching.
-		 */
 		if ( $branch_switch ) {
-			if ( ! empty( $this->type->enterprise_api ) ) {
-				$endpoint = add_query_arg( 'at', $branch_switch, $endpoint );
-			} else {
-				$endpoint = $branch_switch . '.zip';
-			}
+			$endpoint = urldecode( add_query_arg( 'at', $branch_switch, $endpoint ) );
 		}
 
 		return $download_link_base . $endpoint;
+	}
+
+	/**
+	 * Create Bitbucket Server API endpoints.
+	 *
+	 * @param $git      object
+	 * @param $endpoint string
+	 *
+	 * @return string
+	 */
+	protected function add_endpoints( $git, $endpoint ) {
+		switch ( self::$method ) {
+			case 'meta':
+			case 'tags':
+			case 'translation':
+				break;
+			case 'file':
+			case 'readme':
+				$endpoint = add_query_arg( 'at', $git->type->branch, $endpoint );
+				break;
+			case 'changes':
+				$endpoint = add_query_arg( array( 'at' => $git->type->branch, 'raw' => '' ), $endpoint );
+				break;
+			case 'download_link':
+				/*
+				 * Add a prefix query argument to create a subdirectory with the same name
+				 * as the repo, e.g. 'my-repo' becomes 'my-repo/'
+				 * Required for using stash-archive.
+				 */
+				$defaults = array( 'prefix' => $git->type->repo . '/', 'at' => $git->type->branch );
+				$endpoint = add_query_arg( $defaults, $endpoint );
+				if ( ! empty( $git->type->tags ) ) {
+					$endpoint = urldecode( add_query_arg( 'at', $git->type->newest_tag, $endpoint ) );
+				}
+				break;
+			default:
+				break;
+		}
+
+		return $endpoint;
 	}
 
 	/**
@@ -399,134 +429,47 @@ class Bitbucket_API extends API {
 	}
 
 	/**
-	 * Add Basic Authentication $args to http_request_args filter hook
-	 * for private Bitbucket repositories only.
+	 * The Bitbucket Server REST API does not support downloading files directly at the moment
+	 * therefore we'll use this to construct urls to fetch the raw files ourselves.
 	 *
-	 * @param  $args
-	 * @param  $url
+	 * @param string $file filename
 	 *
-	 * @return mixed $args
-	 */
-	public function maybe_authenticate_http( $args, $url ) {
-		$headers       = parse_url( $url );
-		$bitbucket_org = 'bitbucket.org' === $headers['host'] ? true : false;
+	 * @return bool|array false upon failure || return wp_safe_remote_get() response array
+	 **/
+	private function bbenterprise_fetch_raw_file( $file ) {
+		$file         = urlencode( $file );
+		$download_url = '/1.0/projects/:owner/repos/:repo/browse/' . $file;
+		$download_url = $this->add_endpoints( $this, $download_url );
+		$download_url = $this->get_api_url( $download_url );
 
-		$bitbucket_private         = false;
-		$bitbucket_private_install = false;
+		$response = wp_safe_remote_get( $download_url );
 
-		/*
-		 * Check whether attempting to update private Bitbucket repo.
-		 */
-		if ( isset( $this->type->repo ) &&
-		     ! empty( parent::$options[ $this->type->repo ] ) &&
-		     false !== strpos( $url, $this->type->repo )
-		) {
-			$bitbucket_private = true;
+		if ( is_wp_error( $response ) ) {
+			return false;
 		}
 
-		/*
-		 * Check whether attempting to install private Bitbucket repo
-		 * and abort if Bitbucket user/pass not set.
-		 */
-		if ( isset( $_POST['option_page'], $_POST['is_private'] ) &&
-		     'github_updater_install' === $_POST['option_page'] &&
-		     'bitbucket' === $_POST['github_updater_api'] &&
-		     ( ( ! empty( parent::$options['bitbucket_username'] ) &&
-		         ! empty( parent::$options['bitbucket_password'] ) ) ||
-		       ( ! empty( parent::$options['bitbucket_enterprise_username'] ) &&
-		         ! empty( parent::$options['bitbucket_enterprise_password'] ) ) )
-		) {
-			$bitbucket_private_install = true;
-		}
-
-		if ( $bitbucket_private || $bitbucket_private_install ) {
-			$username = $bitbucket_org ? parent::$options['bitbucket_username'] : parent::$options['bitbucket_enterprise_username'];
-			$password = $bitbucket_org ? parent::$options['bitbucket_password'] : parent::$options['bitbucket_enterprise_password'];
-
-			$args['headers']['Authorization'] = 'Basic ' . base64_encode( "$username:$password" );
-		}
-
-		return $args;
+		return $response;
 	}
 
 	/**
-	 * Removes Basic Authentication header for Bitbucket Release Assets.
-	 * Storage in AmazonS3 buckets, uses Query String Request Authentication Alternative.
+	 * Combines separate text lines from API response into one string with \n line endings.
+	 * Code relying on raw text can now parse it.
 	 *
-	 * @link http://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html#RESTAuthenticationQueryStringAuth
+	 * @param string $response
 	 *
-	 * @param $args
-	 * @param $url
-	 *
-	 * @return mixed
+	 * @return string Combined lines of text returned by API
 	 */
-	public function http_release_asset_auth( $args, $url ) {
-		$arrURL = parse_url( $url );
-		if ( isset( $arrURL['host'] ) && 'bbuseruploads.s3.amazonaws.com' === $arrURL['host'] ) {
-			unset( $args['headers']['Authorization'] );
+	private function bbenterprise_recombine_response( $response ) {
+		$remote_info_file = '';
+		$json_decoded     = is_string( $response ) ? json_decode( $response ) : null;
+		$response         = empty( $json_decoded ) ? $response : $json_decoded;
+		if ( isset( $response->lines ) ) {
+			foreach ( $response->lines as $line ) {
+				$remote_info_file .= $line->text . "\n";
+			}
 		}
 
-		return $args;
-	}
-
-	/**
-	 * Added due to abstract class designation, not used for Bitbucket.
-	 *
-	 * @param $git
-	 * @param $endpoint
-	 */
-	protected function add_endpoints( $git, $endpoint ) {
-	}
-
-	/**
-	 * Add Basic Authentication $args to http_request_args filter hook
-	 * for private Bitbucket repositories only during AJAX.
-	 *
-	 * @param $args
-	 * @param $url
-	 *
-	 * @return mixed
-	 */
-	public function ajax_maybe_authenticate_http( $args, $url ) {
-		global $wp_current_filter;
-
-		$headers       = parse_url( $url );
-		$bitbucket_org = 'bitbucket.org' === $headers['host'] ? true : false;
-
-		$ajax_update    = array( 'wp_ajax_update-plugin', 'wp_ajax_update-theme' );
-		$is_ajax_update = array_intersect( $ajax_update, $wp_current_filter );
-
-		if ( ! empty( $is_ajax_update ) ) {
-			$this->load_options();
-		}
-
-		if ( parent::is_doing_ajax() && ! parent::is_heartbeat() &&
-		     ( isset( $_POST['slug'] ) && array_key_exists( $_POST['slug'], parent::$options ) &&
-		       1 == parent::$options[ $_POST['slug'] ] &&
-		       false !== stristr( $url, $_POST['slug'] ) )
-		) {
-			$username = $bitbucket_org ? parent::$options['bitbucket_username'] : parent::$options['bitbucket_enterprise_username'];
-			$password = $bitbucket_org ? parent::$options['bitbucket_password'] : parent::$options['bitbucket_enterprise_password'];
-
-			$args['headers']['Authorization'] = 'Basic ' . base64_encode( "$username:$password" );
-		}
-
-		return $args;
-	}
-
-	/**
-	 * Parse API response call and return only array of tag numbers.
-	 *
-	 * @param object $response Response from API call.
-	 *
-	 * @return array|object Array of tag numbers, object is error.
-	 */
-	protected function parse_tag_response( $response ) {
-		if ( isset( $response->message ) ) {
-			return $response;
-		}
-
-		return array_keys( (array) $response );
+		return $remote_info_file;
 	}
 
 	/**
@@ -541,8 +484,8 @@ class Bitbucket_API extends API {
 		$response = array( $response );
 
 		array_filter( $response, function( $e ) use ( &$arr ) {
-			$arr['private']      = $e->is_private;
-			$arr['last_updated'] = $e->updated_on;
+			$arr['private']      = ! $e->public;
+			$arr['last_updated'] = null;
 			$arr['watchers']     = 0;
 			$arr['forks']        = 0;
 			$arr['open_issues']  = 0;
@@ -552,25 +495,28 @@ class Bitbucket_API extends API {
 	}
 
 	/**
-	 * Parse API response and return array with changelog in base64.
+	 * Parse API response and return array with changelog.
 	 *
 	 * @param object $response Response from API call.
 	 *
-	 * @return array|object $arr Array of changes in base64, object if error.
+	 * @return array $arr Array of changes in base64.
 	 */
 	protected function parse_changelog_response( $response ) {
-		if ( isset( $response->message ) ) {
-			return $response;
-		}
-
-		$arr      = array();
-		$response = array( $response );
-
-		array_filter( $response, function( $e ) use ( &$arr ) {
-			$arr['changes'] = $e->data;
-		} );
-
-		return $arr;
+		return array( 'changes' => $this->bbenterprise_recombine_response( $response ) );
 	}
 
+	/**
+	 * Parse API response and return object with readme body.
+	 *
+	 * @param string $response
+	 *
+	 * @return object $response
+	 */
+	protected function parse_readme_response( $response ) {
+		$content        = $this->bbenterprise_recombine_response( $response );
+		$response       = new \stdClass();
+		$response->data = $content;
+
+		return $response;
+	}
 }