+++ src/GitHub_Updater/GitHub_API.php
@@ -28,6 +28,13 @@ if ( ! defined( 'WPINC' ) ) {
 class GitHub_API extends API {
 
 	/**
+	 * Holds loose class method name.
+	 *
+	 * @var null
+	 */
+	private static $method = null;
+
+	/**
 	 * Constructor.
 	 *
 	 * @param object $type
@@ -48,7 +55,8 @@ class GitHub_API extends API {
 		$response = isset( $this->response[ $file ] ) ? $this->response[ $file ] : false;
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo/contents/' . $file );
+			self::$method = 'file';
+			$response     = $this->api( '/repos/:owner/:repo/contents/' . $file );
 			if ( ! isset( $response->content ) ) {
 				return false;
 			}
@@ -80,7 +88,8 @@ class GitHub_API extends API {
 		$response  = isset( $this->response['tags'] ) ? $this->response['tags'] : false;
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo/tags' );
+			self::$method = 'tags';
+			$response     = $this->api( '/repos/:owner/:repo/tags' );
 
 			if ( ! $response ) {
 				$response          = new \stdClass();
@@ -127,7 +136,8 @@ class GitHub_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo/contents/' . $changes );
+			self::$method = 'changes';
+			$response     = $this->api( '/repos/:owner/:repo/contents/' . $changes );
 
 			if ( $response ) {
 				$response = $this->parse_changelog_response( $response );
@@ -153,9 +163,7 @@ class GitHub_API extends API {
 	 * @return bool
 	 */
 	public function get_remote_readme() {
-		if ( ! file_exists( $this->type->local_path . 'readme.txt' ) &&
-		     ! file_exists( $this->type->local_path_extended . 'readme.txt' )
-		) {
+		if ( ! $this->exists_local_file( 'readme.txt' ) ) {
 			return false;
 		}
 
@@ -175,7 +183,8 @@ class GitHub_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo/contents/readme.txt' );
+			self::$method = 'readme';
+			$response     = $this->api( '/repos/:owner/:repo/contents/readme.txt' );
 		}
 
 		if ( $response && isset( $response->content ) ) {
@@ -203,7 +212,8 @@ class GitHub_API extends API {
 		$response = isset( $this->response['meta'] ) ? $this->response['meta'] : false;
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo' );
+			self::$method = 'meta';
+			$response     = $this->api( '/repos/:owner/:repo' );
 
 			if ( $response ) {
 				$response = $this->parse_meta_response( $response );
@@ -235,7 +245,8 @@ class GitHub_API extends API {
 		}
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/:owner/:repo/branches' );
+			self::$method = 'branches';
+			$response     = $this->api( '/repos/:owner/:repo/branches' );
 
 			if ( $response ) {
 				foreach ( $response as $branch ) {
@@ -338,7 +349,8 @@ class GitHub_API extends API {
 		$type     = explode( '_', $this->type->type );
 
 		if ( ! $response ) {
-			$response = $this->api( '/repos/' . $headers['owner'] . '/' . $headers['repo'] . '/contents/language-pack.json' );
+			self::$method = 'translation';
+			$response     = $this->api( '/repos/' . $headers['owner'] . '/' . $headers['repo'] . '/contents/language-pack.json' );
 
 			if ( $this->validate_response( $response ) ) {
 				return false;
@@ -411,27 +423,25 @@ class GitHub_API extends API {
 	 * @return string $endpoint
 	 */
 	protected function add_endpoints( $git, $endpoint ) {
-
-		/*
-		 * If a branch has been given, only check that for the remote info.
-		 * If it's not been given, GitHub will use the Default branch.
-		 */
-		if ( ! empty( $git->type->branch ) ) {
-			$endpoint = add_query_arg( 'ref', $git->type->branch, $endpoint );
+		switch ( self::$method ) {
+			case 'file':
+			case 'readme':
+				$endpoint = add_query_arg( 'ref', $git->type->branch, $endpoint );
+				break;
+			case 'meta':
+			case 'tags':
+			case 'changes':
+			case 'download_link':
+			case 'translation':
+				break;
+			default:
+				break;
 		}
 
 		$endpoint = $this->add_access_token_endpoint( $git, $endpoint );
 
 		/*
-		 * Remove branch endpoint if a translation file.
-		 */
-		$repo = explode( '/', $endpoint );
-		if ( isset( $repo[3] ) && $repo[3] !== $git->type->repo ) {
-			$endpoint = remove_query_arg( 'ref', $endpoint );
-		}
-
-		/*
-		 * If using GitHub Enterprise header return this endpoint.
+		 * If GitHub Enterprise return this endpoint.
 		 */
 		if ( ! empty( $git->type->enterprise_api ) ) {
 			return $git->type->enterprise_api . $endpoint;
@@ -592,4 +602,5 @@ class GitHub_API extends API {
 
 		return $args;
 	}
+
 }