+++ src/GitHub_Updater/Rest_Update.php
@@ -343,6 +343,8 @@ class Rest_Update extends Base {
 
 		//$response['payload'] = $new;
 
+		//$response['payload'] = $new;
+
 		return $response;
 	}