+++ CHANGES.md
@@ -4,6 +4,9 @@
 * added a singleton to `class Settings` to avoid duplicate loads [#531](https://github.com/afragen/github-updater/issues/531)
 * refactored subtabs for Settings page
 * refactored parsing of extra headers, `Enterprise` and `CE` headers no longer needed
+* added support for Bitbucket Server!! Thanks @lkistenkas for access and especially to @BjornW for kicking it off
+* refactored `add_endpoints()` to use everywhere
+* now requires WordPress 4.4 and above
 
 #### 6.2.2 / 2017-02-09
 * fixed for updating via webhook from GitHub tagged release, declare branch as `master`